backend/requirements.txt
fastapi
uvicorn
pydantic
jose
python-jose[cryptography]
python-multipart
bcrypt
motor
aiosmtplib
python-dotenv
aiosmtplib
cloudinary
pydantic[email]